+++ src/miner/miner.cpp
@@ -39,8 +39,9 @@ void miner(uint32_t core)
         {
             break;
         }
-
+#if defined(ESP32)
         current_update_hashrate();
+#endif
     }
 
 #if defined(HAS_LCD)
@@ -49,6 +50,10 @@ void miner(uint32_t core)
 
     l_info(TAG_MINER, "[%d] > [%s] > 0x%.8x - diff %.12f", core, current_job->job_id.c_str(), winning_nonce, diff_hash);
     network_send(current_job->job_id, current_job->extranonce2, current_job->ntime, winning_nonce);
+
+#if defined(ESP8266)
+    current_update_hashrate();
+#endif
 
     current_setHighestDifficulty(diff_hash);