+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # v0.0.15
 
 - Fix setup.html adding auto update
-- Fix 
+- Fix configuration dump
+- Tweak ESP8266 hash rate print
+- Add 0xe000 for esp32 for boot